Fix typo in package name + style
