+++ legacy-fabric-entity-events-v1/1.8.9/src/main/java/net/legacyfabric/fabric/mixin/entity/event/versioned/PlayerManagerMixin.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package net.legacyfabric.fabric.mixin.entity.event.versionned;
+package net.legacyfabric.fabric.mixin.entity.event.versioned;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
